requirements.txt
beautifulsoup4
lxml
pyperclip

requests
flask
environs
python-dotenv
xmltodict